pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "company-memory-agent"
version = "0.1.0"
description = "Python CLI agent that ingests company info into a two-layer JSON memory and answers questions via LLM-routed retrieval."
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "trafilatura>=1.12",
    "pypdf>=4.2",
    "python-docx>=1.1",
    "pydantic>=2.6",
    "typer>=0.12",
    "reportlab>=4.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
agent = "agent.cli:app"

[tool.setuptools.packages.find]
include = ["agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
